GPX support

+++ mediaserver/domain/lap.go
@@ -5,8 +5,8 @@ import "time"
 type Lap struct {
 	StartTimestamp           time.Time `json:"startTimestamp"`
 	EndTimestamp             time.Time `json:"endTimestamp"`
-	DistanceInLapMetres      int       `json:"distanceInLapMetres"`
-	CumulativeDistanceMetres int       `json:"cumulativeDistanceMetres"`
-	StartAltitude            int       `json:"startAltitude"`
-	EndAltitude              int       `json:"endAltitude"`
+	DistanceInLapMetres      float64   `json:"distanceInLapMetres"`
+	CumulativeDistanceMetres float64   `json:"cumulativeDistanceMetres"`
+	StartAltitude            float64   `json:"startAltitude"`
+	EndAltitude              float64   `json:"endAltitude"`
 }
